refactor: Move CreateNoteRequest to fixtures
